Prints fixes and gpu Power read updates

+++ goamdsmi_shim/rsmi/rocm_smi_go_shim.c
@@ -75,7 +75,7 @@ goamdsmi_status_t go_shim_rsmi_num_monitor_devices(uint32_t* gpu_num_monitor_dev
     if(RSMI_STATUS_SUCCESS == rsmi_num_monitor_devices(&gpu_num_monitor_devices_temp))
     {
         *gpu_num_monitor_devices = gpu_num_monitor_devices_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success, GpuNumMonitorDevices:%lu\n", *gpu_num_monitor_devices);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success, GpuNumMonitorDevices:%lu\n", (unsigned long)(*gpu_num_monitor_devices));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -123,7 +123,7 @@ goamdsmi_status_t go_shim_rsmi_dev_pci_id_get(uint32_t dv_ind, uint64_t* gpu_pci
     if(RSMI_STATUS_SUCCESS == rsmi_dev_pci_id_get(dv_ind, &gpu_pci_id_temp))
     {
         *gpu_pci_id = gpu_pci_id_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuPciId:%llu\n", dv_ind, (*gpu_pci_id));}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuPciId:%llu\n", dv_ind, (unsigned long long)(*gpu_pci_id));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -173,7 +173,7 @@ goamdsmi_status_t go_shim_rsmi_dev_power_cap_get(uint32_t dv_ind, uint64_t* gpu_
     if(RSMI_STATUS_SUCCESS == rsmi_dev_power_cap_get(dv_ind, GPU_SENSOR_0, &gpu_power_cap_temp))
     {
         *gpu_power_cap = gpu_power_cap_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuPowerCap:%llu, GpuPowerCapInWatt:%.6f\n", dv_ind, (*gpu_power_cap), ((double)(*gpu_power_cap))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuPowerCap:%llu, GpuPowerCapInWatt:%.6f\n", dv_ind, (unsigned long long)(*gpu_power_cap), ((double)(*gpu_power_cap))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -188,7 +188,7 @@ goamdsmi_status_t go_shim_rsmi_dev_power_get(uint32_t dv_ind, uint64_t* gpu_powe
     if(RSMI_STATUS_SUCCESS == rsmi_dev_power_ave_get(dv_ind, GPU_SENSOR_0, &gpu_power_temp))
     {
         *gpu_power = gpu_power_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuPowerAverage:%llu, GpuPowerAverageinWatt:%.6f\n", dv_ind, (*gpu_power), ((double)(*gpu_power))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuPowerAverage:%llu, GpuPowerAverageinWatt:%.6f\n", dv_ind, (unsigned long long)(*gpu_power), ((double)(*gpu_power))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -196,12 +196,12 @@ goamdsmi_status_t go_shim_rsmi_dev_power_get(uint32_t dv_ind, uint64_t* gpu_powe
     if(RSMI_STATUS_SUCCESS == rsmi_dev_power_get(dv_ind, &gpu_power_temp, &power_type))
     {
         *gpu_power = gpu_power_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d PowerType:%d, PowerTypeStr:%s, GpuPower:%llu, GpuPowerinWatt:%.6f\n", dv_ind, (int)power_type, powerTypesToString[power_type], (*gpu_power), ((double)(*gpu_power))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d PowerType:%d, PowerTypeStr:%s, GpuPower:%llu, GpuPowerinWatt:%.6f\n", dv_ind, (int)power_type, powerTypesToString[power_type], (unsigned long long)(*gpu_power), ((double)(*gpu_power))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
     else
     {
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Failed for Gpu:%d PowerType:%d, PowerTypeStr:%s, GpuPower:%llu, GpuPowerinWatt:%.6f\n", dv_ind, (int)power_type, powerTypesToString[power_type], (*gpu_power), ((double)(*gpu_power))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Failed for Gpu:%d PowerType:%d, PowerTypeStr:%s, GpuPower:%llu, GpuPowerinWatt:%.6f\n", dv_ind, (int)power_type, powerTypesToString[power_type], (unsigned long long)(*gpu_power), ((double)(*gpu_power))/1000000);}
     }
 
     return GOAMDSMI_STATUS_FAILURE;
@@ -215,7 +215,7 @@ goamdsmi_status_t go_shim_rsmi_dev_temp_metric_get(uint32_t dv_ind, uint32_t sen
     if(RSMI_STATUS_SUCCESS == rsmi_dev_temp_metric_get(dv_ind, sensor, metric, &gpu_temperature_temp))
     {
         *gpu_temperature = gpu_temperature_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d Sensor:%d Metric:%d, GpuTemperature:%llu, GpuTemperatureInDegree:%.3f\n", dv_ind, sensor, metric, (*gpu_temperature), ((double)(*gpu_temperature))/1000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d Sensor:%d Metric:%d, GpuTemperature:%llu, GpuTemperatureInDegree:%.3f\n", dv_ind, sensor, metric, (unsigned long long)(*gpu_temperature), ((double)(*gpu_temperature))/1000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -230,7 +230,7 @@ goamdsmi_status_t go_shim_rsmi_dev_overdrive_level_get(uint32_t dv_ind, uint32_t
     if(RSMI_STATUS_SUCCESS == rsmi_dev_overdrive_level_get(dv_ind, &gpu_overdrive_level_temp))
     {
         *gpu_overdrive_level = gpu_overdrive_level_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuOverdriveLevel:%lu\n", dv_ind, *gpu_overdrive_level);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuOverdriveLevel:%lu\n", dv_ind, (unsigned long)(*gpu_overdrive_level));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -245,7 +245,7 @@ goamdsmi_status_t go_shim_rsmi_dev_mem_overdrive_level_get(uint32_t dv_ind, uint
     if(RSMI_STATUS_SUCCESS == rsmi_dev_mem_overdrive_level_get(dv_ind, &gpu_mem_overdrive_level_temp))
     {
         *gpu_mem_overdrive_level = gpu_mem_overdrive_level_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemoryOverdriveLevel:%lu\n", dv_ind, *gpu_mem_overdrive_level);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemoryOverdriveLevel:%lu\n", dv_ind, (unsigned long)(*gpu_mem_overdrive_level));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -260,7 +260,7 @@ goamdsmi_status_t go_shim_rsmi_dev_perf_level_get(uint32_t dv_ind, uint32_t *gpu
     if(RSMI_STATUS_SUCCESS == rsmi_dev_perf_level_get(dv_ind, &perf))
     {
         *gpu_perf = perf;                //Conversion from rsmi_dev_perf_level_t to uint32_t
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuPerfLevel:%lu\n", dv_ind, *gpu_perf);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuPerfLevel:%lu\n", dv_ind, (unsigned long)(*gpu_perf));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -275,7 +275,7 @@ goamdsmi_status_t go_shim_rsmi_dev_gpu_clk_freq_get_sclk(uint32_t dv_ind, uint64
     if(RSMI_STATUS_SUCCESS == rsmi_dev_gpu_clk_freq_get(dv_ind, RSMI_CLK_TYPE_SYS, &freq))
     {
         *gpu_sclk_freq = freq.frequency[freq.current];
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuSclkFreq:%llu, GpuSclkFreqMhz:%.6f\n", dv_ind, (*gpu_sclk_freq), ((double)(*gpu_sclk_freq))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuSclkFreq:%llu, GpuSclkFreqMhz:%.6f\n", dv_ind, (unsigned long long)(*gpu_sclk_freq), ((double)(*gpu_sclk_freq))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -290,7 +290,7 @@ goamdsmi_status_t go_shim_rsmi_dev_gpu_clk_freq_get_mclk(uint32_t dv_ind, uint64
     if(RSMI_STATUS_SUCCESS == rsmi_dev_gpu_clk_freq_get(dv_ind, RSMI_CLK_TYPE_MEM, &freq))
     {
         *gpu_memclk_freq = freq.frequency[freq.current];
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMclkFreq:%llu, GpuMclkFreqMhz:%.6f\n", dv_ind, (*gpu_memclk_freq), ((double)(*gpu_memclk_freq))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMclkFreq:%llu, GpuMclkFreqMhz:%.6f\n", dv_ind, (unsigned long long)(*gpu_memclk_freq), ((double)(*gpu_memclk_freq))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -305,7 +305,7 @@ goamdsmi_status_t go_shim_rsmi_od_volt_freq_range_min_get_sclk(uint32_t dv_ind, 
     if(RSMI_STATUS_SUCCESS == rsmi_dev_od_volt_info_get(dv_ind, &odv))
     {
         *gpu_min_sclk = odv.curr_sclk_range.lower_bound;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuSclkMinfreq:%llu, GpuSclkMinfreqMhz:%.6f\n", dv_ind, (*gpu_min_sclk), ((double)(*gpu_min_sclk))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuSclkMinfreq:%llu, GpuSclkMinfreqMhz:%.6f\n", dv_ind, (unsigned long long)(*gpu_min_sclk), ((double)(*gpu_min_sclk))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -320,7 +320,7 @@ goamdsmi_status_t go_shim_rsmi_od_volt_freq_range_min_get_mclk(uint32_t dv_ind, 
     if(RSMI_STATUS_SUCCESS == rsmi_dev_od_volt_info_get(dv_ind, &odv))
     {
         *gpu_min_memclk = odv.curr_mclk_range.lower_bound;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemclkMinfreq:%llu, GpuMemclkMinfreqMhz:%.6f\n", dv_ind, (*gpu_min_memclk), ((double)(*gpu_min_memclk))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemclkMinfreq:%llu, GpuMemclkMinfreqMhz:%.6f\n", dv_ind, (unsigned long long)(*gpu_min_memclk), ((double)(*gpu_min_memclk))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -335,7 +335,7 @@ goamdsmi_status_t go_shim_rsmi_od_volt_freq_range_max_get_sclk(uint32_t dv_ind, 
     if(RSMI_STATUS_SUCCESS == rsmi_dev_od_volt_info_get(dv_ind, &odv))
     {
         *gpu_max_sclk = odv.curr_sclk_range.upper_bound;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuSclkMaxfreq:%llu, GpuSclkMaxfreqMhz:%.6f\n", dv_ind, (*gpu_max_sclk), ((double)(*gpu_max_sclk))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuSclkMaxfreq:%llu, GpuSclkMaxfreqMhz:%.6f\n", dv_ind, (unsigned long long)(*gpu_max_sclk), ((double)(*gpu_max_sclk))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -350,7 +350,7 @@ goamdsmi_status_t go_shim_rsmi_od_volt_freq_range_max_get_mclk(uint32_t dv_ind, 
     if(RSMI_STATUS_SUCCESS == rsmi_dev_od_volt_info_get(dv_ind, &odv))
     {
         *gpu_max_memclk = odv.curr_mclk_range.upper_bound;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemclkMaxfreq:%llu, GpuMemclkMaxfreqMhz:%.6f\n", dv_ind, (*gpu_max_memclk), ((double)(*gpu_max_memclk))/1000000);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemclkMaxfreq:%llu, GpuMemclkMaxfreqMhz:%.6f\n", dv_ind, (unsigned long long)(*gpu_max_memclk), ((double)(*gpu_max_memclk))/1000000);}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -365,7 +365,7 @@ goamdsmi_status_t go_shim_rsmi_dev_gpu_busy_percent_get(uint32_t dv_ind, uint32_
     if(RSMI_STATUS_SUCCESS == rsmi_dev_busy_percent_get(dv_ind, &gpu_busy_percent_temp))
     {
         *gpu_busy_percent = gpu_busy_percent_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuBusyPerc:%lu\n", dv_ind, *gpu_busy_percent);}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuBusyPerc:%lu\n", dv_ind, (unsigned long)(*gpu_busy_percent));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -382,7 +382,7 @@ goamdsmi_status_t go_shim_rsmi_dev_gpu_memory_busy_percent_get(uint32_t dv_ind, 
         (RSMI_STATUS_SUCCESS == rsmi_dev_memory_total_get(dv_ind, RSMI_MEM_TYPE_VRAM, &gpu_memory_total_temp)))
     {
         *gpu_memory_busy_percent = (uint64_t)(gpu_memory_usage_temp*100)/gpu_memory_total_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemoryBusyPerc:%llu\n", dv_ind, (*gpu_memory_busy_percent));}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemoryBusyPerc:%llu\n", dv_ind, (unsigned long long)(*gpu_memory_busy_percent));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -397,7 +397,7 @@ goamdsmi_status_t go_shim_rsmi_dev_gpu_memory_usage_get(uint32_t dv_ind, uint64_
     if(RSMI_STATUS_SUCCESS == rsmi_dev_memory_usage_get(dv_ind, RSMI_MEM_TYPE_VRAM, &gpu_memory_usage_temp))
     {
         *gpu_memory_usage = (uint64_t)gpu_memory_usage_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemoryUsage:%llu\n", dv_ind, (*gpu_memory_usage));}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemoryUsage:%llu\n", dv_ind, (unsigned long long)(*gpu_memory_usage));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
 
@@ -412,7 +412,7 @@ goamdsmi_status_t go_shim_rsmi_dev_gpu_memory_total_get(uint32_t dv_ind, uint64_
     if(RSMI_STATUS_SUCCESS == rsmi_dev_memory_total_get(dv_ind, RSMI_MEM_TYPE_VRAM, &gpu_memory_total_temp))
     {
         *gpu_memory_total = (uint64_t)gpu_memory_total_temp;
-        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemoryTotal:%llu\n", dv_ind, (*gpu_memory_total));}
+        if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {printf("ROCMSMI, Success for Gpu:%d, GpuMemoryTotal:%llu\n", dv_ind, (unsigned long long)(*gpu_memory_total));}
         return GOAMDSMI_STATUS_SUCCESS;
     }
